split all grape related functions to separate file  - casuing problems in pytest so I want to isolate them  - require a separate workflow so it makes sense
